ai-workflow-builder/server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
chromadb==0.4.18
openai==1.3.8
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
serpapi==0.1.5
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.3
tiktoken==0.5.2